Refactored the end2end tests
 -created 2 new files and renamed the original file so it is
 more clear for other developers that want to read the code or
 use it for their own purposes
